requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.5
env==0.1.0
environs==9.5.0
gunicorn==21.2.0
marshmallow==3.19.0
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
slugify==0.0.1
sqlparse==0.4.3
